refactor(magnetohydrodynamics): replace full-solver task with cpp-full1d-00
