--- benchmarks/magnetohydrodynamics/cpp-full-solver1d/workspace/README.md
@@ -1,3 +1,0 @@
-The public C++ workspace now contains the hardcoded Brio-Wu solver.
-
-Shared workspace docs are already mounted for this benchmark.
