+++ benchmarks/magnetohydrodynamics/cpp-full1d-00/eval/run.sh
