+++ benchmarks/magnetohydrodynamics/cpp-full1d-00/task.toml
@@ -1,8 +1,8 @@
-id = "cpp-full-solver1d"
+id = "cpp-full1d-00"
 suite = "magnetohydrodynamics"
 language = "cpp"
 time_limit_sec = 600
 eval_cmd = "/eval/run.sh"
-prompt = "Read /run/spec.md and solve the task in /work."
+prompt = "Read /run/spec.md, /work/basic_equations.md, and /work/hlld.md, then solve the task in /work."
 use_shared_workspace = true
 use_shared_eval = true
