+++ benchmarks/magnetohydrodynamics/cpp-full1d-00/workspace/README.md
@@ -1,0 +1,7 @@
+The public C++ workspace contains a Brio-Wu solver scaffold.
+
+- `src/main.cpp` is complete.
+- `src/hlld.cpp` is complete.
+- `src/mhd1d.cpp` contains TODO sections to implement.
+
+Shared workspace docs are already mounted for this benchmark.
