+++ benchmarks/magnetohydrodynamics/cpp-full1d-00/workspace/pyproject.toml
@@ -1,5 +1,5 @@
 [project]
-name = "magnetohydrodynamics-cpp-full-solver1d"
+name = "magnetohydrodynamics-cpp-full1d-00"
 version = "0.0.0"
 requires-python = ">=3.10"
