+++ benchmarks/magnetohydrodynamics/cpp-full1d-00/workspace/src/hlld.hpp
@@ -1,0 +1,4 @@
+#pragma once
+
+void hlld_flux_from_primitive(const double* left, const double* right, double bx, double gamma,
+                              double* flux);
