+++ benchmarks/magnetohydrodynamics/cpp-full1d-00/workspace/src/main.cpp
@@ -5,9 +5,9 @@
 #include <iostream>
 #include <stdexcept>
 
-constexpr int                DefaultNx = 100;
-constexpr double             Gamma     = 2.0;
-constexpr double             Bx        = 0.75;
+constexpr int                Nx    = 100;
+constexpr double             Gamma = 2.0;
+constexpr double             Bx    = 0.75;
 constexpr mhd1d::StateVector LeftPrimitive{
     1.0, 0.0, 0.0, 0.0, 1.0, 1.0, 0.0,
 };
@@ -18,7 +18,7 @@ constexpr mhd1d::StateVector RightPrimitive{
 int parse_nx(int argc, char** argv)
 {
   if (argc <= 1) {
-    return DefaultNx;
+    return Nx;
   }
 
   char*      end    = nullptr;
@@ -43,7 +43,7 @@ mhd1d::SolverWorkspace initialize(int nx, double gamma, double bx,
   }
 
   mhd1d::set_boundary(workspace.up, workspace.up, workspace.Lbx, workspace.Ubx);
-  mhd1d::primitive_profile_to_conservative(workspace.up, workspace.uc, bx, gamma);
+  mhd1d::convert_primitive_to_conservative(workspace.up, workspace.uc, bx, gamma);
 
   return workspace;
 }
