+++ benchmarks/magnetohydrodynamics/cpp-full1d-00/workspace/tests/data/brio_wu_golden.csv
