+++ benchmarks/magnetohydrodynamics/shared/src/mhd1d.cpp
@@ -1,4 +1,4 @@
-#include "full_mhd1d.hpp"
+#include "mhd1d.hpp"
 #include "hlld.hpp"
 
 #include <algorithm>
