+++ benchmarks/magnetohydrodynamics/shared/src/mhd1d.hpp
